requirements.txt
py-clob-client>=0.17.0
websocket-client>=1.7.0
pandas>=2.2.0
numpy>=1.26.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
aiohttp>=3.9.0
loguru>=0.7.2
